load stuff from config
