--- bot/bot_test.go
@@ -1,9 +1,0 @@
-package bot
-
-/*
-Why doesnt this work?
-func TestCanAddHandler(t *testing.T) {
-	b := NewDiscordBot("a token")
-	b.Session.AddHandler(RunTestCommand)
-}
-*/
