+++ bot_test.go
@@ -1,14 +1,22 @@
-package config
+package main
 
 import (
 	"strings"
 	"testing"
 )
 
+/*
+Why doesnt this work?
+func TestCanAddHandler(t *testing.T) {
+	b := NewDiscordBot("a token")
+	b.Session.AddHandler(RunTestCommand)
+}
+*/
+
 // Tests that the Load method is able to load valid JSON data from a reader into a struct
 func TestLoadConfigValidData(t *testing.T) {
 	reader := strings.NewReader(`{"Token" : "atoken", "DbConnStr" : "dbconnstr"}`)
-	b := &Bot{}
+	b := &DiscordBot{}
 	err := b.Load(reader)
 	if err != nil {
 		t.Errorf(err.Error())
@@ -26,7 +34,7 @@ func TestLoadConfigValidData(t *testing.T) {
 // Tests that the Load method is able to load valid JSON data from a reader into a struct
 func TestLoadConfigMissingDataCausesInvalid(t *testing.T) {
 	reader := strings.NewReader(`{"DbConnStr" : "dbconnstr"}`)
-	b := &Bot{}
+	b := &DiscordBot{}
 	err := b.Load(reader)
 	if err != nil {
 		t.Errorf(err.Error())
