+++ db.go
@@ -1,4 +1,4 @@
-package db
+package main
 
 import (
 	"context"
